+++ AmplitudeTests/AmplitudeTests.m
@@ -16,8 +16,13 @@
 #import "AMPDeviceInfo.h"
 #import "AMPARCMacros.h"
 #import "AMPUtils.h"
+#import "AMPTrackingOptions.h"
 
 // expose private methods for unit testing
+@interface AMPDeviceInfo (Tests)
++(NSString*)getAdvertiserID:(int) maxAttempts;
+@end
+
 @interface Amplitude (Tests)
 - (NSDictionary*)mergeEventsAndIdentifys:(NSMutableArray*)events identifys:(NSMutableArray*)identifys numEvents:(long) numEvents;
 - (id) truncate:(id) obj;
@@ -815,6 +820,68 @@
     XCTAssertNotEqualObjects(oldDeviceId, newDeviceId);
     XCTAssertEqualObjects(newDeviceId, [dbHelper getValue:@"device_id"]);
     XCTAssertTrue([newDeviceId hasSuffix:@"R"]);
+}
+
+-(void)testTrackIdfa {
+    id mockDeviceInfo = OCMClassMock([AMPDeviceInfo class]);
+    [[mockDeviceInfo expect] getAdvertiserID:5];
+
+    Amplitude *client = [Amplitude instanceWithName:@"has_idfa"];
+    [client flushQueueWithQueue:client.initializerQueue];
+    [client initializeApiKey:@"blah"];
+    [client flushQueue];
+
+    [client logEvent:@"test"];
+    [client flushQueue];
+
+    [mockDeviceInfo verify];
+    [mockDeviceInfo stopMocking];
+}
+
+-(void)testDisableIdfa {
+    id mockDeviceInfo = OCMClassMock([AMPDeviceInfo class]);
+    [[mockDeviceInfo reject] getAdvertiserID:5];
+
+    Amplitude *client = [Amplitude instanceWithName:@"disable_idfa"];
+    [client flushQueueWithQueue:client.initializerQueue];
+    [client disableIdfaTracking];
+    [client initializeApiKey:@"blah"];
+    [client flushQueue];
+
+    [client logEvent:@"test"];
+    [client flushQueue];
+
+    [mockDeviceInfo verify];
+    [mockDeviceInfo stopMocking];
+}
+
+-(void)testSetTrackingConfig {
+    AMPTrackingOptions *options = [[[[[AMPTrackingOptions options] disableCity] disableIPAddress] disableLanguage] disableCountry];
+    [self.amplitude setTrackingOptions:options];
+
+    [self.amplitude logEvent:@"test"];
+    [self.amplitude flushQueue];
+    NSDictionary *event = [self.amplitude getLastEvent];
+
+    // verify we have platform and carrier since those were not filtered out
+    XCTAssertEqualObjects([event objectForKey:@"platform"], @"iOS");
+    XCTAssertEqualObjects([event objectForKey:@"carrier"], @"Unknown");
+
+    // verify we do not have any of the filtered out events
+    XCTAssertNil([event objectForKey:@"city"]);
+    XCTAssertNil([event objectForKey:@"country"]);
+    XCTAssertNil([event objectForKey:@"language"]);
+
+    // verify api properties contains tracking options for location filtering
+    NSDictionary *apiProperties = [event objectForKey:@"api_properties"];
+    XCTAssertNotNil([apiProperties objectForKey:@"ios_idfv"]);
+    XCTAssertNotNil([apiProperties objectForKey:@"tracking_options"]);
+
+    NSDictionary *trackingOptions = [apiProperties objectForKey:@"tracking_options"];
+    XCTAssertEqual(3, trackingOptions.count);
+    XCTAssertEqualObjects([NSNumber numberWithBool:NO], [trackingOptions objectForKey:@"city"]);
+    XCTAssertEqualObjects([NSNumber numberWithBool:NO], [trackingOptions objectForKey:@"country"]);
+    XCTAssertEqualObjects([NSNumber numberWithBool:NO], [trackingOptions objectForKey:@"ip_address"]);
 }
 
 @end